refactor(app): share the import logic between server and demo
The demo's import repeated the server's: creating the run, persisting
executions, linking their files, rolling a trace group's counters up,
recognising a repeat upload, shaping the response. That is the duplication
AGENTS.md rules out, and it was already drifting — the two had different
duplicate checks for the same situation.

It moves to `#shared/handlers/import-runs.ts`, which both call. Only four
things genuinely differ between the runtimes, and they arrive as an
`ImportPort`: how executions are persisted (each side already had its own),
where files are stored (blob store versus IndexedDB), how a trace's console is
read out of a nested archive (node zlib versus DecompressionStream), and how
the dashboard is told a run appeared. The pre-flight's verdict rules are shared
the same way, with only the size ceiling differing.

What is left in each caller is what should be: the server endpoint is auth,
size limits and multipart parsing; the demo handler is form reading and the
browser's ZIP. The server endpoint drops from 711 lines to 250, its pre-flight
from 133 to 88, and the demo's from 450 to 200.

Behaviour is unchanged on both sides: the same 12 end-to-end tests pass, and
the demo still imports a real blob report into IndexedDB with its 7 executions
and 11 files, served back through the service worker.
